sky_power_estimation/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sky_power_estimation"
version = "1.0.0"
description = "Multi-modal deep learning for DC power estimation from sky images"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Sky Power Estimation Team", email = "team@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["solar", "forecasting", "deep-learning", "computer-vision", "power-prediction"]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "timm>=0.9.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "pvlib>=0.10.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
full = [
    "albumentations>=1.3.0",
    "tensorboard>=2.14.0",
    "wandb>=0.15.0",
    "ephem>=4.1.0",
    "seaborn>=0.12.0",
]

[project.scripts]
sky-power-train = "scripts.train:main"
sky-power-predict = "scripts.predict:main"
sky-power-evaluate = "scripts.evaluate:main"

[project.urls]
Homepage = "https://github.com/example/sky-power-estimation"
Documentation = "https://github.com/example/sky-power-estimation#readme"
Repository = "https://github.com/example/sky-power-estimation"

[tool.setuptools.packages.find]
where = ["."]
include = ["sky_power_estimation*", "scripts*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=sky_power_estimation"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
